requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
pillow
tqdm
sqlite-utils
